refactor: :truck: Move heroes_controller to data folder.
