+++ lib/data/heroes_controller.dart
@@ -1,8 +1,8 @@
-import '../data/api_service/data_api.dart';
-import '../data/dao_service/shared_prefs.dart';
 import '../domain/models/marvel_hero.dart';
+import 'api_service/data_api.dart';
+import 'dao_service/shared_prefs.dart';
 
-class HeroesRepository {//TODO: MOVER PARA DATA
+class HeroesRepository {
   final DataApi _dataApi = DataApi();
   final SharedPrefs _prefs = SharedPrefs();
