+++ lib/domain/business/bloc/home_bloc/home_bloc.dart
@@ -1,7 +1,7 @@
 import 'package:dio/dio.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 
-import '../../../../controllers/heroes_controller.dart';
+import '../../../../data/heroes_controller.dart';
 import 'home_events.dart';
 import 'home_states.dart';
